notary/requirements.txt
astroid==2.3.3
attrdict==2.0.1
attrs==19.3.0
base58==0.2.2
boto3==1.11.9
botocore==1.14.9
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cobra-hdwallet==0.3.1
cytoolz==0.10.1
docutils==0.15.2
ecdsa==0.13
eth-abi==2.1.0
eth-account==0.4.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.8.4
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-PyMongo==2.3.0
hexbytes==0.2.0
idna==2.8
importlib-metadata==1.5.0
ipfshttpclient==0.4.12
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.4
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lru-dict==1.1.6
MarkupSafe==1.1.1
mccabe==0.6.1
mnemonic==0.13
multiaddr==0.0.9
netaddr==0.7.19
parsimonious==0.8.1
pbkdf2==1.3
protobuf==3.11.3
pycryptodome==3.9.6
PyJWT==1.7.1
pylint==2.4.4
pymongo==3.10.1
pyrsistent==0.15.7
pysha3==1.0.2
python-dateutil==2.8.1
python-dotenv==0.10.5
requests==2.22.0
rlp==1.2.0
rope==0.16.0
s3transfer==0.3.2
six==1.14.0
toolz==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.8
varint==1.0.2
web3==5.5.0
websockets==8.1
Werkzeug==0.16.1
wrapt==1.11.2
zipp==2.1.0
